Introduce views generator
This introduces the `suspenders:views` generator, responsible for
populating the Rails app with some useful base templates.

One of the templates brings in the `analytics` partial, which does not
exist until Segment is installed. When that is extracted we should make
sure to handle this.
