+++ lib/suspenders.rb
@@ -10,6 +10,7 @@ require "suspenders/generators/factories_generator"
 require "suspenders/generators/lint_generator"
 require "suspenders/generators/jobs_generator"
 require "suspenders/generators/analytics_generator"
+require "suspenders/generators/views_generator"
 require "suspenders/actions"
 require "suspenders/adapters/heroku"
 require "suspenders/app_builder"
