+++ lib/suspenders/app_builder.rb
@@ -153,32 +153,6 @@ config.public_file_server.headers = {
       remove_file "config/initializers/wrap_parameters.rb"
     end
 
-    def create_partials_directory
-      empty_directory 'app/views/application'
-    end
-
-    def create_shared_flashes
-      copy_file "_flashes.html.erb", "app/views/application/_flashes.html.erb"
-      copy_file "flashes_helper.rb", "app/helpers/flashes_helper.rb"
-    end
-
-    def create_shared_javascripts
-      copy_file '_javascript.html.erb', 'app/views/application/_javascript.html.erb'
-    end
-
-    def create_shared_css_overrides
-      copy_file(
-        "_css_overrides.html.erb",
-        "app/views/application/_css_overrides.html.erb",
-      )
-    end
-
-    def create_application_layout
-      template 'suspenders_layout.html.erb.erb',
-        'app/views/layouts/application.html.erb',
-        force: true
-    end
-
     def use_postgres_config_template
       template 'postgresql_database.yml.erb', 'config/database.yml',
         force: true
