+++ lib/suspenders/generators/app_generator.rb
@@ -46,7 +46,6 @@ module Suspenders
       invoke :setup_test_environment
       invoke :setup_production_environment
       invoke :setup_secret_token
-      invoke :create_suspenders_views
       invoke :configure_app
       invoke :copy_miscellaneous_files
       invoke :customize_error_pages
@@ -116,15 +115,6 @@ module Suspenders
     def setup_secret_token
       say 'Moving secret token out of version control'
       build :setup_secret_token
-    end
-
-    def create_suspenders_views
-      say 'Creating suspenders views'
-      build :create_partials_directory
-      build :create_shared_flashes
-      build :create_shared_javascripts
-      build :create_shared_css_overrides
-      build :create_application_layout
     end
 
     def configure_app
@@ -213,6 +203,7 @@ module Suspenders
       generate("suspenders:lint")
       generate("suspenders:jobs")
       generate("suspenders:analytics")
+      generate("suspenders:views")
     end
 
     def outro
